+++ src/Api/Providers/Pinners.php
@@ -4,6 +4,7 @@ namespace seregazhuk\PinterestBot\Api\Providers;
 
 use Iterator;
 use seregazhuk\PinterestBot\Api\Response;
+use seregazhuk\PinterestBot\Helpers\Pagination;
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
 use seregazhuk\PinterestBot\Api\Traits\Followable;
 use seregazhuk\PinterestBot\Api\Traits\Searchable;
@@ -13,10 +14,13 @@ class Pinners extends Provider
 {
     use Searchable, Followable;
 
+    /**
+     * @var array
+     */
     protected $loginRequiredFor = [
         'follow',
-        'unFollow',
         'block',
+        'unFollow',
         'blockById',
     ];
 
@@ -51,7 +55,7 @@ class Pinners extends Provider
      * @return Iterator
      * @throws WrongFollowingType
      */
-    public function following($username, $type = UrlBuilder::FOLLOWING_PEOPLE, $limit = 0)
+    public function following($username, $type = UrlBuilder::FOLLOWING_PEOPLE, $limit = Pagination::DEFAULT_LIMIT)
     {
         $followingUrl = UrlBuilder::getFollowingUrlByType($type);
 
@@ -70,7 +74,7 @@ class Pinners extends Provider
      * @param int $limit
      * @return Iterator
      */
-    public function followingPeople($username, $limit = 0)
+    public function followingPeople($username, $limit = Pagination::DEFAULT_LIMIT)
     {
         return $this->following($username, UrlBuilder::FOLLOWING_PEOPLE, $limit);
     }
@@ -83,7 +87,7 @@ class Pinners extends Provider
      * @param int $limit
      * @return Iterator
      */
-    public function followingBoards($username, $limit = 0)
+    public function followingBoards($username, $limit = Pagination::DEFAULT_LIMIT)
     {
         return $this->following($username, UrlBuilder::FOLLOWING_BOARDS, $limit);
     }
@@ -97,7 +101,7 @@ class Pinners extends Provider
      * @return Iterator
      * @throws WrongFollowingType
      */
-    public function followingInterests($username, $limit = 0)
+    public function followingInterests($username, $limit = Pagination::DEFAULT_LIMIT)
     {
         return $this->following($username, UrlBuilder::FOLLOWING_INTERESTS, $limit);
     }
@@ -110,7 +114,7 @@ class Pinners extends Provider
      *
      * @return Iterator
      */
-    public function pins($username, $limit = 0)
+    public function pins($username, $limit = Pagination::DEFAULT_LIMIT)
     {
         return $this->paginate(
             $username, UrlBuilder::RESOURCE_USER_PINS, $limit
@@ -124,7 +128,7 @@ class Pinners extends Provider
      * @param int $limit
      * @return Iterator
      */
-    public function likes($username, $limit = 0)
+    public function likes($username, $limit = Pagination::DEFAULT_LIMIT)
     {
         return $this->paginate(
             $username, UrlBuilder::RESOURCE_USER_LIKES, $limit
@@ -161,15 +165,13 @@ class Pinners extends Provider
      * @param string $url
      * @param int $limit
      *
-     * @return Iterator
+     * @return \Traversable
      */
     protected function paginate($username, $url, $limit)
     {
-        $params = [
-            'data' => ['username' => $username],
-            'url'  => $url,
-        ];
-
-        return $this->getPaginatedResponse($params, $limit);
+        return (new Pagination($limit))
+            ->paginateOver(function($bookmarks = []) use ($username, $url) {
+                return $this->getPaginatedData(['username' => $username], $url, $bookmarks);
+            });
     }
 }